Move text copying to MessageList. Fix some edge cases.
