+++ src/main.js
@@ -27,7 +27,6 @@ import GlobalApi from '@/libs/GlobalApi';
 import { AudioManager } from '@/libs/AudioManager';
 import { SoundBleep } from '@/libs/SoundBleep';
 import WindowTitle from '@/libs/WindowTitle';
-import CopyTextHandler from '@/libs/CopyTextHandler';
 import { configTemplates } from '@/res/configTemplates';
 
 // Global utilities
@@ -426,7 +425,6 @@ function initInputCommands() {
 
 function startApp() {
     new WindowTitle(state);
-    new CopyTextHandler(state);
 
     api.emit('init');
